+++ sdk/src/main/java/com/bugsnag/android/Bugsnag.java
@@ -19,7 +19,8 @@ import java.util.Map;
 @SuppressWarnings("checkstyle:JavadocTagContinuationIndentation")
 public final class Bugsnag {
 
-    @Nullable
+    private static final Object lock = new Object();
+
     @SuppressLint("StaticFieldLeak")
     static Client client;
 
@@ -33,9 +34,7 @@ public final class Bugsnag {
      */
     @NonNull
     public static Client init(@NonNull Context androidContext) {
-        client = new Client(androidContext);
-        NativeInterface.configureClientObservers(client);
-        return client;
+        return init(androidContext, null, true);
     }
 
     /**
@@ -46,9 +45,7 @@ public final class Bugsnag {
      */
     @NonNull
     public static Client init(@NonNull Context androidContext, @Nullable String apiKey) {
-        client = new Client(androidContext, apiKey);
-        NativeInterface.configureClientObservers(client);
-        return client;
+        return init(androidContext, apiKey, true);
     }
 
     /**
@@ -62,9 +59,9 @@ public final class Bugsnag {
     public static Client init(@NonNull Context androidContext,
                               @Nullable String apiKey,
                               boolean enableExceptionHandler) {
-        client = new Client(androidContext, apiKey, enableExceptionHandler);
-        NativeInterface.configureClientObservers(client);
-        return client;
+        Configuration config
+            = ConfigFactory.createNewConfiguration(androidContext, apiKey, enableExceptionHandler);
+        return init(androidContext, config);
     }
 
     /**
@@ -75,9 +72,21 @@ public final class Bugsnag {
      */
     @NonNull
     public static Client init(@NonNull Context androidContext, @NonNull Configuration config) {
-        client = new Client(androidContext, config);
-        NativeInterface.configureClientObservers(client);
+        synchronized (lock) {
+            if (client == null) {
+                client = new Client(androidContext, config);
+                NativeInterface.configureClientObservers(client);
+            } else {
+                logClientInitWarning();
+            }
+        }
         return client;
+    }
+
+    private static void logClientInitWarning() {
+        Logger.warn("It appears that Bugsnag.init() was called more than once. Subsequent "
+            + "calls have no effect, but may indicate that Bugsnag is not integrated in an"
+            + " Application subclass, which can lead to undesired behaviour.");
     }
 
     /**